Move environment config for toggles into a separate file (doesn't need to be a directory)
